Updates to remaps for Ulauncher
Talked to a Ulauncher dev and determined that the Ulauncher search results are limited for the foreseeable future to a max of 17 items. So the screen size is not relevant and there is no reason at present to remap beyond Ctrl+h. So this keymap block can be trimmed down a bit. 

Remapping Ctrl+a and Ctrl+c of course interferes with quickly selecting and copying text within the search field, so it seems wise to disable those by default, and the trimmed list will no longer have Ctrl+v as a remap, so it won't interfere with pasting text into the search field. I don't see any good way around these conflicts, except maybe shifting the whole remap scheme to using Ctrl+Shift instead of just the Ctrl key. But that wouldn't be very intuitive. 

Most desired results will tend to be near the top of the results list anyway, so the primary benefit is still retained by having 1-9 mapped, plus the remaining letters. 

Adding `KEY_0` to the numerical key remaps, which will have no effect for now since Alt+0 doesn't exist, but I'll be asking the Ulauncher devs to add it as a shortcut in the app, so I hope it will be a usable remap in a future Ulauncher release.